chore(app): rename typings to fix file errors
